refactor Layout into Header and Footer
